+++ .teamcity/settings.kts
@@ -43,7 +43,7 @@ project {
 
 
 object UnitTests : BuildType({
-    val gradleVersions = listOf("6.7.1", "6.9.2", "7.5.1")
+    val gradleVersions = listOf("6.8", "6.9.2", "7.5.1")
 
     name = "Unit Tests"