+++ src/main/kotlin/org/jetbrains/intellij/IntelliJPlugin.kt
@@ -86,7 +86,7 @@ open class IntelliJPlugin : Plugin<Project> {
                 // IdeaModel.project is available only for root project
                 this.project?.settings {
                     taskTriggers {
-                        afterSync("setupDependencies")
+                        afterSync(IntelliJPluginConstants.SETUP_DEPENDENCIES_TASK_NAME)
                     }
                 }
             }
@@ -123,8 +123,8 @@ open class IntelliJPlugin : Plugin<Project> {
     }
 
     private fun checkGradleVersion(project: Project) {
-        if (Version.parse(project.gradle.gradleVersion) < Version.parse("6.7.1")) {
-            throw PluginInstantiationException("${IntelliJPluginConstants.NAME} requires Gradle 6.7.1 and higher")
+        if (Version.parse(project.gradle.gradleVersion) < Version.parse("6.8")) {
+            throw PluginInstantiationException("${IntelliJPluginConstants.NAME} requires Gradle 6.8 and higher")
         }
     }