+++ src/test/kotlin/org/jetbrains/intellij/IntelliJPluginSpec.kt
@@ -481,18 +481,18 @@ class IntelliJPluginSpec : IntelliJPluginSpecBase() {
     }
 
     @Test
-    fun `throws exception if Gradle is lt 6_7_1`() {
-        val message = "Gradle IntelliJ Plugin requires Gradle 6.7.1 and higher"
+    fun `throws exception if Gradle is lt 6_8`() {
+        val message = "Gradle IntelliJ Plugin requires Gradle 6.8 and higher"
 
         build("6.4", true, "help").output.let {
             assertTrue(it.contains("FAILURE: Build failed with an exception."))
         }
 
-        build("6.7", true, "help").output.let {
+        build("6.7.1", true, "help").output.let {
             assertTrue(it.contains("FAILURE: Build failed with an exception."))
         }
 
-        build("6.7.1", false, "help").output.let {
+        build("6.8", false, "help").output.let {
             assertTrue(it.contains("BUILD SUCCESSFUL"))
             assertFalse(it.contains(message))
         }