+++ src/test/kotlin/org/jetbrains/intellij/tasks/JarSearchableOptionsTaskSpec.kt
@@ -9,6 +9,7 @@ import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
 
+@Suppress("ComplexRedundantLet")
 class JarSearchableOptionsTaskSpec : SearchableOptionsSpecBase() {
 
     @Test